refactor: DAO 클래스 추상화
