+++ src/test/java/kr/co/_29cm/homework/dao/InMemoryOrderDaoTest.java
@@ -6,16 +6,16 @@ import kr.co._29cm.homework.domain.Order;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
-public class OrderDaoTest {
+public class InMemoryOrderDaoTest {
 
     @DisplayName("주문을 저장한다.")
     @Test
     void save() {
         // given
-        final OrderDao orderDao = new OrderDao();
+        final InMemoryOrderDao inMemoryOrderDao = new InMemoryOrderDao();
         final Order order = new Order(5000, 0);
 
         // when & then
-        assertDoesNotThrow(() -> orderDao.save(order));
+        assertDoesNotThrow(() -> inMemoryOrderDao.save(order));
     }
 }
