+++ src/test/java/kr/co/_29cm/homework/dao/InMemoryOrderProductDaoTest.java
@@ -6,16 +6,16 @@ import kr.co._29cm.homework.domain.OrderProduct;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
-public class OrderProductDaoTest {
+public class InMemoryOrderProductDaoTest {
 
     @DisplayName("주문 상품을 저장한다.")
     @Test
     void save() {
         // given
-        final OrderProductDao orderProductDao = new OrderProductDao();
+        final InMemoryOrderProductDao inMemoryOrderProductDao = new InMemoryOrderProductDao();
         final OrderProduct orderProduct = new OrderProduct(1L, 1L, 10);
 
         // when & then
-        assertDoesNotThrow(() -> orderProductDao.save(orderProduct));
+        assertDoesNotThrow(() -> inMemoryOrderProductDao.save(orderProduct));
     }
 }
